Reorganization of everything, which causes tons of renames
Code now lives in src/tsgl, tests in src/tsgl/tests.cpp.
Build is in bin/ build/ and lib/
Data was moved to assets/
Librarys to build against are still in libraries/, but still gitignored as we are
trying to make people use their own.
Added makefile and first version of configure