Fixed docs and tried to make screenshot faster
makefile: Added optimisation and debugging flags. Moved include
and bad dependency to let 'make all' build
Canvas.cpp: Moved the mutex lock in drawPoint() to make it faster
Removed stringstream in screenShot(), replacing with sprintf() as
it is much simpler and faster.
Timer.h: Added better documentation for getTimeBetweenSleeps()
tests.cpp: Rewrote getPixelsFunction to make it faster and easier
to see the threads working. Removed the screen recording from it.